Update Test files
Test file separated.
