+++ src/Task.test.js
@@ -78,56 +78,4 @@ describe('Tasks', () => {
       expect(tasks[0].iscomplete).toBe(false);
     });
   });
-  describe('ClearCompletedTasks', () => {
-    it('should remove all completed tasks from the list', () => {
-      const taskso = new Tasks();
-
-      // Add some tasks to the list, with one completed
-      taskso.AddTask('Task 1');
-      taskso.AddTask('Task 2');
-      taskso.AddTask('Task 3');
-      taskso.CheckTask(1, true); // Mark Task 1 as completed
-
-      // Run ClearCompletedTasks() to remove completed tasks
-      taskso.ClearCompletedTasks();
-      const tasks = JSON.parse(localStorage.getItem('tasks'));
-
-      // Ensure that completed Task 1 is removed, and the other tasks remain
-      expect(tasks).toHaveLength(2);
-      expect(tasks.map(task => task.description)).toEqual(['Task 2', 'Task 3']);
-    });
-
-    it('should not remove any tasks if no tasks are completed', () => {
-      const taskso = new Tasks();
-
-      // Add some tasks to the list, none completed
-      taskso.AddTask('Task 1');
-      taskso.AddTask('Task 2');
-
-      // Run ClearCompletedTasks() when no tasks are completed
-      taskso.ClearCompletedTasks();
-      const tasks = JSON.parse(localStorage.getItem('tasks'));
-
-      // Ensure that all tasks remain unchanged
-      expect(tasks).toHaveLength(2);
-      expect(tasks.map(task => task.description)).toEqual(['Task 1', 'Task 2']);
-    });
-
-    it('should not remove any tasks if the list is empty', () => {
-      const taskso = new Tasks();
-
-      // Run ClearCompletedTasks() when the list is empty
-      taskso.ClearCompletedTasks();
-      const tasks = JSON.parse(localStorage.getItem('tasks'));
-
-      // Ensure that the list remains empty
-      expect(tasks).toHaveLength(0);
-    });
-  });
-
-  // Cleanup after all tests
-  afterAll(() => {
-    // Remove the list container element mock from the DOM
-    document.body.removeChild(listContainerMock);
-  });
 });
